+++ abecto-benchmark/src/main/java/de/uni_jena/cs/fusion/abecto/benchmark/PopulationComparisonProcessorBenchmark.java
@@ -17,9 +17,14 @@ package de.uni_jena.cs.fusion.abecto.benchmark;
 
 import de.uni_jena.cs.fusion.abecto.Aspect;
 import de.uni_jena.cs.fusion.abecto.processor.PopulationComparisonProcessor;
+import org.apache.jena.query.Query;
+import org.apache.jena.query.QueryFactory;
+import org.apache.jena.rdf.model.Model;
 import org.apache.jena.rdf.model.Resource;
-import org.openjdk.jmh.annotations.Benchmark;
+import org.apache.jena.rdf.model.ResourceFactory;
+import org.openjdk.jmh.annotations.*;
 
+import java.util.Collections;
 import java.util.List;
 import java.util.Set;
 import java.util.stream.Stream;
@@ -27,9 +32,30 @@ import java.util.stream.Stream;
 public class PopulationComparisonProcessorBenchmark {
 
     @Benchmark
-    public void testMethod() {
-        // This is a demo/sample template for building your JMH benchmarks. Edit as needed.
-        // Put your benchmark code here.
+    @Fork(value = 1, warmups = 2)
+    @Warmup(iterations = 5)
+    @Measurement(iterations = 10)
+    @BenchmarkMode(Mode.SingleShotTime)
+    public Model benchmark(Configuration configuration) {
+        PopulationComparisonProcessor processor = configuration.processor;
+        processor.run();
+        return processor.getOutputMetaModel(ResourceFactory.createResource("0"));
+    }
+
+    @State(Scope.Benchmark)
+    public static class Configuration {
+        @Param({"100", "1000", "10000", "100000"})
+        public int populationSize;
+        @Param({"2", "3", "4", "5", "6", "7", "8", "9", "10"})
+        public int datasetCount;
+        @Param({"0.25", "0.5", "0.75"})
+        public double coverage;
+        public PopulationComparisonProcessor processor;
+
+        @Setup(Level.Iteration)
+        public void setup() {
+            processor = new PopulationComparisonProcessorBenchmark.IndependentPopulationComparisonProcessor(this.populationSize, this.datasetCount, this.coverage);
+        }
     }
 
     private static class IndependentPopulationComparisonProcessor extends PopulationComparisonProcessor {
@@ -38,6 +64,14 @@ public class PopulationComparisonProcessorBenchmark {
 
         public IndependentPopulationComparisonProcessor(int populationSize, int datasetCount, double coverage) {
             this.dataSupplier = new ComparisonBenchmarkDataSupplier(populationSize, datasetCount, coverage, 0);
+            Resource aspectIri = ResourceFactory.createResource("aspect");
+            this.aspects = Collections.singleton(aspectIri);
+            Aspect aspect = new Aspect(aspectIri, "key");
+            Query query = QueryFactory.create("SELECT ?var1 WHERE {?key ?p ?var1}");
+            for (Resource dataset : dataSupplier.getDatasets()) {
+                aspect.setPattern(dataset, query);
+            }
+            this.addAspects(aspect);
         }
 
         @Override