+++ abecto-benchmark/src/main/java/de/uni_jena/cs/fusion/abecto/benchmark/PropertyComparisonProcessorBenchmark.java
@@ -17,9 +17,13 @@ package de.uni_jena.cs.fusion.abecto.benchmark;
 
 import de.uni_jena.cs.fusion.abecto.Aspect;
 import de.uni_jena.cs.fusion.abecto.processor.PropertyComparisonProcessor;
+import org.apache.jena.query.Query;
+import org.apache.jena.query.QueryFactory;
+import org.apache.jena.rdf.model.Model;
 import org.apache.jena.rdf.model.RDFNode;
 import org.apache.jena.rdf.model.Resource;
-import org.openjdk.jmh.annotations.Benchmark;
+import org.apache.jena.rdf.model.ResourceFactory;
+import org.openjdk.jmh.annotations.*;
 
 import java.util.*;
 import java.util.stream.Stream;
@@ -27,17 +31,48 @@ import java.util.stream.Stream;
 public class PropertyComparisonProcessorBenchmark {
 
     @Benchmark
-    public void testMethod() {
-        // This is a demo/sample template for building your JMH benchmarks. Edit as needed.
-        // Put your benchmark code here.
+    @Fork(value = 1, warmups = 2)
+    @Warmup(iterations = 5)
+    @Measurement(iterations = 10)
+    @BenchmarkMode(Mode.SingleShotTime)
+    public Model benchmark(Configuration configuration) {
+        PropertyComparisonProcessor processor = configuration.processor;
+        processor.run();
+        return processor.getOutputMetaModel(ResourceFactory.createResource("0"));
+    }
+
+    @State(Scope.Benchmark)
+    public static class Configuration {
+        @Param({"100", "1000", "10000", "100000"})
+        public int populationSize;
+        @Param({"2", "3", "4", "5", "6", "7", "8", "9", "10"})
+        public int datasetCount;
+        @Param({"0.25", "0.5", "0.75"})
+        public double coverage;
+        @Param({"0.1"})
+        public double errorRate;
+        public PropertyComparisonProcessor processor;
+
+        @Setup(Level.Iteration)
+        public void setup() {
+            processor = new IndependentPropertyComparisonProcessor(this.populationSize, this.datasetCount, this.coverage, this.errorRate);
+        }
     }
 
     private static class IndependentPropertyComparisonProcessor extends PropertyComparisonProcessor {
 
         private final ComparisonBenchmarkDataSupplier dataSupplier;
 
-        public IndependentPropertyComparisonProcessor(int populationSize, int datasetCount, double coverage) {
-            this.dataSupplier = new ComparisonBenchmarkDataSupplier(populationSize, datasetCount, coverage, 0);
+        public IndependentPropertyComparisonProcessor(int populationSize, int datasetCount, double coverage, double errorRate) {
+            this.dataSupplier = new ComparisonBenchmarkDataSupplier(populationSize, datasetCount, coverage, errorRate);
+            this.aspect = ResourceFactory.createResource("aspect");
+            this.variables = Collections.singletonList("var1");
+            Aspect aspect = new Aspect(this.aspect, "key");
+            Query query = QueryFactory.create("SELECT ?var1 WHERE {?key ?p ?var1}");
+            for (Resource dataset : dataSupplier.getDatasets()) {
+                aspect.setPattern(dataset, query);
+            }
+            this.addAspects(aspect);
         }
 
         @Override